refactor: 파일명 및 확장자 변경
