+++ src/containers/PurchaseResult/PurchaseResult.style.css
