+++ src/containers/WinningResult/WinningResult.style.css
